+++ pkg/pmapi/messages.go
@@ -532,7 +532,13 @@ func (c *Client) GetMessage(id string) (msg *Message, err error) {
 		return
 	}
 
+	if res.StatusCode == http.StatusUnprocessableEntity {
+		err = ErrNoSuchMessage
+		return
+	}
+
 	msg, err = res.Message, res.Err()
+
 	return
 }