+++ Changelog.md
@@ -15,6 +15,7 @@ Changelog [format](http://keepachangelog.com/en/1.0.0/)
 * GODT-282 Completely delete old draft instead moving to trash when user updates draft
 * Adding DSN Sentry as build time parameter
 * GODT-124 bump go-appdir from v1.0.0 to v1.1.0
+* Skip processing message update event if http statuscode is 422
 
 ## [v1.2.6] Donghai - beta (2020-03-XXX)